add VideoMode
